app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
aiohttp>=3.9.0
databricks-sdk>=0.36.0
